proteger rutas con auth

+++ routes/web.php
@@ -16,25 +16,25 @@ Auth::routes();
 
 Route::get('/home', [HomeController::class, 'index'])->name('home');
 
-Route::get('/membresias', [MembresiaController::class, 'index'])->name('membresias.nueva');
-Route::get('/membresias/lista', [MembresiaController::class, 'list'])->name('membresias.list');
-Route::post('/membresias/store', [MembresiaController::class, 'store'])->name('membresias.store');
-Route::put('/membresias/{membresia}', [MembresiaController::class, 'update'])->name('membresias.update');
-Route::delete('/membresias/{id}', [MembresiaController::class, 'destroy'])->name('membresias.destroy');
+Route::get('/membresias', [MembresiaController::class, 'index'])->name('membresias.nueva')->middleware('auth');
+Route::get('/membresias/lista', [MembresiaController::class, 'list'])->name('membresias.list')->middleware('auth');
+Route::post('/membresias/store', [MembresiaController::class, 'store'])->name('membresias.store')->middleware('auth');
+Route::put('/membresias/{membresia}', [MembresiaController::class, 'update'])->name('membresias.update')->middleware('auth');
+Route::delete('/membresias/{id}', [MembresiaController::class, 'destroy'])->name('membresias.destroy')->middleware('auth');
 
 // Rutas para Usuarios
-Route::get('/usuarios', [UserController::class, 'index'])->name('usuarios.nueva');
-Route::get('/usuarios/lista', [UserController::class, 'list'])->name('usuarios.list');
-Route::post('/usuarios/store', [UserController::class, 'store'])->name('usuarios.store');
-Route::put('/usuarios/{usuario}', [UserController::class, 'update'])->name('usuarios.update');
-Route::delete('/usuarios/{id}', [UserController::class, 'destroy'])->name('usuarios.destroy');
+Route::get('/usuarios', [UserController::class, 'index'])->name('usuarios.nueva')->middleware('auth');
+Route::get('/usuarios/lista', [UserController::class, 'list'])->name('usuarios.list')->middleware('auth');
+Route::post('/usuarios/store', [UserController::class, 'store'])->name('usuarios.store')->middleware('auth');
+Route::put('/usuarios/{usuario}', [UserController::class, 'update'])->name('usuarios.update')->middleware('auth');
+Route::delete('/usuarios/{id}', [UserController::class, 'destroy'])->name('usuarios.destroy')->middleware('auth');
 
 
-Route::get('/clases', [ClaseController::class, 'index'])->name('clases.nueva');
-Route::get('/clases/lista', [ClaseController::class, 'list'])->name('clases.list');
-Route::post('/clases/store', [ClaseController::class, 'store'])->name('clases.store');
-Route::put('/clases/{clase}', [ClaseController::class, 'update'])->name('clases.update');
-Route::delete('/clases/{id}', [ClaseController::class, 'destroy'])->name('clases.destroy');
+Route::get('/clases', [ClaseController::class, 'index'])->name('clases.nueva')->middleware('auth');
+Route::get('/clases/lista', [ClaseController::class, 'list'])->name('clases.list')->middleware('auth');
+Route::post('/clases/store', [ClaseController::class, 'store'])->name('clases.store')->middleware('auth');
+Route::put('/clases/{clase}', [ClaseController::class, 'update'])->name('clases.update')->middleware('auth');
+Route::delete('/clases/{id}', [ClaseController::class, 'destroy'])->name('clases.destroy')->middleware('auth');
 
 // Manda a llamar al metodo redirect en los controlador
 Route::get('auth/google/redirect', [GoogleAuthController::class, 'redirect'])->name('auth.google.redirect');
